Rename api.* to rest.*, make immutable properties final
