+++ src/main/java/edu/kit/ipd/crowdcontrol/objectservice/Main.java
@@ -1,6 +1,6 @@
 package edu.kit.ipd.crowdcontrol.objectservice;
 
-import edu.kit.ipd.crowdcontrol.objectservice.api.Router;
+import edu.kit.ipd.crowdcontrol.objectservice.rest.Router;
 
 public class Main {
 	public static void main(String[] args) {
