+++ src/main/java/edu/kit/ipd/crowdcontrol/objectservice/rest/NotAcceptableException.java
@@ -1,4 +1,4 @@
-package edu.kit.ipd.crowdcontrol.objectservice.api;
+package edu.kit.ipd.crowdcontrol.objectservice.rest;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
